// Code/Websites/DanpheEMR/wwwroot/DanpheApp/src/app/settings-new/shared/DTOs/print-templates-settings.dto.ts
export class PrintTemplateSettings_DTO {

    PrintTemplateSettingsId: number = 0;
    PrintType: string = "";
    FieldSettingsName: string = "";
    PrintTemplateMainFormat: string = "";
    PrintTemplateDetailsFormat: string = "";
    PrinterType: string = "";
    VisitType: string = "";
    IsActive: boolean = false;
}